# docs/swagger.yaml
definitions:
...
  ds.ChatResponseWithFlags:
    properties:
      friends:
        type: integer
      id:
        type: integer
      img:
        type: string
      info:
        type: string
      is_read:
        type: boolean
      name:
        type: string
      nickname:
        type: string
      sound:
        type: boolean
      subscribers:
        type: integer
    type: object
  ds.ErrorResp:
    properties:
      description:
        type: string
      status:
        type: string
    type: object
...
  ds.MessageDetail:
    properties:
      chats:
        items:
          $ref: '#/definitions/ds.ChatResponseWithFlags'
        type: array
      creator:
        type: string
      date_create:
        type: string
      date_finish:
        type: string
      date_update:
        type: string
      id:
        type: integer
      moderator:
        type: string
      status:
        type: string
      text:
        type: string
    type: object
  ds.MessageWithUsers:
    properties:
      creator:
        type: string
      date_create:
        type: string
      date_finish:
        type: string
      date_update:
        type: string
      id:
        type: integer
      moderator:
        type: string
      status:
        type: string
      text:
        type: string
    type: object
  ds.OkResp:
    properties:
      message:
        type: string
    type: object
  ds.UpdateMessageTextInput:
    properties:
      text:
        description: Обязательное поле текста
        type: string
    required:
    - text
    type: object
  ds.UpdateMessageTextResp:
    properties:
      text:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /message-chats/delete/{message_id}/{chat_id}:
    delete:
      description: Удаляет чат с указанным ID из сообщения с указанным ID.
      parameters:
      - description: ID сообщения
        in: path
        name: message_id
        required: true
        type: integer
      - description: ID чата
        in: path
        name: chat_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном удалении чата
          schema:
            $ref: '#/definitions/ds.OkResp'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Удаление чата из сообщения
      tags:
      - Message-Chats
  /message-chats/switch/{message_id}/{chat_id}:
    put:
      description: Переключает значение поля "со звуком" у чата в сообщении.
      parameters:
      - description: ID сообщения
        in: path
        name: message_id
        required: true
        type: integer
      - description: ID чата
        in: path
        name: chat_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном изменении значения
          schema:
            $ref: '#/definitions/ds.OkResp'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Переключение поля "со звуком"
      tags:
      - Message-Chats
  /messages:
    get:
      consumes:
      - application/json
      description: Возвращает список сообщений, отфильтрованных по статусу и диапазону
        дат.
      parameters:
      - description: Статус сообщения
        in: query
        name: status
        type: string
      - description: Начальная дата (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Конечная дата (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список сообщений
          schema:
            items:
              $ref: '#/definitions/ds.MessageWithUsers'
            type: array
        "400":
          description: Некорректный формат дат
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Получить отфильтрованные сообщения
      tags:
      - Messages
  /messages/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает полные данные о сообщении, включая чаты.
      parameters:
      - description: ID сообщения
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Детальная информация о сообщении
          schema:
            $ref: '#/definitions/ds.MessageDetail'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Получить сообщение по ID
      tags:
      - Messages
  /messages/{id}/delete:
    delete:
      consumes:
      - application/json
      description: Устанавливает статус сообщения на "удалён".
      parameters:
      - description: ID сообщения
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном обновлении статуса
          schema:
            $ref: '#/definitions/ds.OkResp'
        "403":
          description: Действие запрещено
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "409":
          description: Конфликт статусов
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Удалить сообщение
      tags:
      - Messages
  /messages/{id}/finish:
    put:
      consumes:
      - application/json
      description: Устанавливает статус сообщения на "завершён".
      parameters:
      - description: ID сообщения
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном обновлении статуса
          schema:
            $ref: '#/definitions/ds.OkResp'
        "409":
          description: Конфликт статусов
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Завершить сообщение
      tags:
      - Messages
  /messages/{id}/form:
    put:
      consumes:
      - application/json
      description: Устанавливает статус сообщения на "сформирован".
      parameters:
      - description: ID сообщения
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном обновлении статуса
          schema:
            $ref: '#/definitions/ds.OkResp'
        "400":
          description: Неверный ID сообщения или пустой текст
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "403":
          description: Действие запрещено
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "409":
          description: Конфликт статусов
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Сформировать сообщение
      tags:
      - Messages
  /messages/{id}/reject:
    put:
      consumes:
      - application/json
      description: Устанавливает статус сообщения на "отклонён".
      parameters:
      - description: ID сообщения
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном обновлении статуса
          schema:
            $ref: '#/definitions/ds.OkResp'
        "409":
          description: Конфликт статусов
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Отклонить сообщение
      tags:
      - Messages
  /messages/{id}/text:
    put:
      consumes:
      - application/json
      description: Обновляет текст сообщения по ID.
      parameters:
      - description: ID сообщения
        in: path
        name: id
        required: true
        type: integer
      - description: Новый текст сообщения
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/ds.UpdateMessageTextInput'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый текст
          schema:
            $ref: '#/definitions/ds.UpdateMessageTextResp'
        "400":
          description: Некорректный ID или неверный формат данных
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Обновить текст сообщения
      tags:
      - Messages
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"